+++ docker/base/debian-7/conf/bin/service.d/supervisor.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+exec supervisord -c /opt/docker/etc/supervisor.conf --logfile /dev/null --pidfile /dev/null --user root