+++ docker/base/ubuntu-15.10/conf/bin/config.sh
@@ -224,5 +224,5 @@ function runDockerProvision() {
  ##
 function startSupervisord() {
     cd /
-    exec supervisord -c /opt/docker/etc/supervisor.conf --logfile /dev/null --pidfile /dev/null --user root
+    exec /opt/docker/bin/service.d/supervisor.sh
 }