+++ test/run.sh
@@ -200,6 +200,9 @@ initEnvironment
     OS_VERSION="8" runTestForTag "debian-8"
     OS_VERSION="testing" runTestForTag "debian-9"
 
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
+
     waitForTestRun
 }
 
@@ -223,6 +226,9 @@ initEnvironment
     OS_VERSION="7" runTestForTag "debian-7"
     OS_VERSION="8" runTestForTag "debian-8"
     OS_VERSION="testing" runTestForTag "debian-9"
+
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
 
     waitForTestRun
 }
@@ -248,6 +254,37 @@ initEnvironment
     OS_VERSION="8" runTestForTag "debian-8"
     OS_VERSION="testing" runTestForTag "debian-9"
 
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
+
+    waitForTestRun
+}
+
+
+#######################################
+# webdevops/base
+#######################################
+
+[[ $(checkTestTarget base-app) ]] && {
+    setupTestEnvironment "base-app"
+
+    OS_VERSION="12.04" runTestForTag "ubuntu-12.04"
+    OS_VERSION="14.04" runTestForTag "ubuntu-14.04"
+    OS_VERSION="15.04" runTestForTag "ubuntu-15.04"
+    OS_VERSION="15.10" runTestForTag "ubuntu-15.10"
+    OS_VERSION="16.04" runTestForTag "ubuntu-16.04"
+
+    setEnvironmentOsFamily "redhat"
+    OS_VERSION="7" runTestForTag "centos-7"
+
+    setEnvironmentOsFamily "debian"
+    OS_VERSION="7" OS_VERSION="7" runTestForTag "debian-7"
+    OS_VERSION="8" runTestForTag "debian-8"
+    OS_VERSION="testing" runTestForTag "debian-9"
+
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
+
     waitForTestRun
 }
 
@@ -259,30 +296,8 @@ initEnvironment
     setupTestEnvironment "php"
     setSpecTest "php5"
 
-    OS_VERSION="12.04" runTestForTag "ubuntu-12.04"
-    OS_VERSION="14.04" runTestForTag "ubuntu-14.04"
-    OS_VERSION="15.04" runTestForTag "ubuntu-15.04"
-    OS_VERSION="15.10" runTestForTag "ubuntu-15.10"
-    OS_VERSION="16.04" runTestForTag "ubuntu-16.04"
-
-    waitForTestRun
-
-    setEnvironmentOsFamily "redhat"
-    OS_VERSION="7" runTestForTag "centos-7"
-
-    setEnvironmentOsFamily "debian"
-    OS_VERSION="7" runTestForTag "debian-7"
-    OS_VERSION="8" runTestForTag "debian-8"
-    OS_VERSION="testing" runTestForTag "debian-9"
-
-    setEnvironmentOsFamily "ubuntu"
-    setSpecTest "php7"
-    OS_VERSION="16.04" runTestForTag "ubuntu-16.04-php7"
-
-    setEnvironmentOsFamily "debian"
-    setSpecTest "php7"
-    OS_VERSION="8" runTestForTag "debian-8-php7"
-    OS_VERSION="testing" runTestForTag "debian-9-php7"
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
 
     waitForTestRun
 }
@@ -308,6 +323,9 @@ initEnvironment
     OS_VERSION="8" runTestForTag "debian-8"
     OS_VERSION="testing" runTestForTag "debian-9"
 
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
+
     waitForTestRun
 }
 
@@ -332,6 +350,9 @@ initEnvironment
     OS_VERSION="8" runTestForTag "debian-8"
     OS_VERSION="testing" runTestForTag "debian-9"
 
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
+
     waitForTestRun
 }
 
@@ -349,8 +370,6 @@ initEnvironment
     OS_VERSION="15.10" runTestForTag "ubuntu-15.10"
     OS_VERSION="16.04" runTestForTag "ubuntu-16.04"
 
-    waitForTestRun
-
     setEnvironmentOsFamily "redhat"
     OS_VERSION="7" runTestForTag "centos-7"
 
@@ -358,6 +377,9 @@ initEnvironment
     OS_VERSION="7" runTestForTag "debian-7"
     OS_VERSION="8" runTestForTag "debian-8"
     OS_VERSION="testing" runTestForTag "debian-9"
+
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
 
     setEnvironmentOsFamily "ubuntu"
     setSpecTest "php7-apache"
@@ -385,8 +407,6 @@ initEnvironment
     OS_VERSION="15.10" runTestForTag "ubuntu-15.10"
     OS_VERSION="16.04" runTestForTag "ubuntu-16.04"
 
-    waitForTestRun
-
     setEnvironmentOsFamily "redhat"
     OS_VERSION="7" runTestForTag "centos-7"
 
@@ -394,6 +414,9 @@ initEnvironment
     OS_VERSION="7" runTestForTag "debian-7"
     OS_VERSION="8" runTestForTag "debian-8"
     OS_VERSION="testing" runTestForTag "debian-9"
+
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "alpine-3"
 
     setEnvironmentOsFamily "ubuntu"
     setSpecTest "php7-nginx"
@@ -481,6 +504,18 @@ initEnvironment
 [[ $(checkTestTarget ssh) ]] && {
     setupTestEnvironment "ssh"
     OS_VERSION="$DOCKER_TAG_LATEST" runTestForTag "latest"
+
+    waitForTestRun
+}
+
+#######################################
+# webdevops/sphinx
+#######################################
+
+[[ $(checkTestTarget sphinx) ]] && {
+    setupTestEnvironment "sphinx"
+    setEnvironmentOsFamily "alpine"
+    OS_VERSION="3" runTestForTag "latest"
 
     waitForTestRun
 }