+++ test/spec/docker/php5_spec.rb
@@ -10,6 +10,7 @@ describe "Dockerfile" do
 
     include_examples 'collection::bootstrap'
     include_examples 'collection::base'
+    include_examples 'collection::base-app'
     include_examples 'collection::php5'
     include_examples 'collection::php-fpm5'
     include_examples 'collection::php-fpm5::public'