+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [WebDevOps.io Dockerfile](https://github.com/webdevops/Dockerfile).
 
+## [0.22.0] - upcoming
+### Changed
+- Introduced base-app for application images
+- Added alpine-3 images
+
 ## [0.21.6] - 2016-03-25
 ### Changed
 - Refactored directory layout (container -> docker/, .bin -> bin, testsuite -> test/